% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_graph.R
\name{add_knowledge_graph_edge}
\alias{add_knowledge_graph_edge}
\title{Add an edge to a knowledge graph specification}
\usage{
add_knowledge_graph_edge(graph, ...)
}
\arguments{
\item{graph}{Knowledge graph specification.}

\item{...}{Arguments passed to \code{\link[=knowledge_graph_edge]{knowledge_graph_edge()}}.}
}
\value{
Updated \code{agentr_knowledge_graph_spec}.
}
\description{
Add an edge to a knowledge graph specification
}
