% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_graph.R
\name{add_knowledge_graph_node}
\alias{add_knowledge_graph_node}
\title{Add a node to a knowledge graph specification}
\usage{
add_knowledge_graph_node(graph, ...)
}
\arguments{
\item{graph}{Knowledge graph specification.}

\item{...}{Arguments passed to \code{\link[=knowledge_graph_node]{knowledge_graph_node()}}.}
}
\value{
Updated \code{agentr_knowledge_graph_spec}.
}
\description{
Add a node to a knowledge graph specification
}
