% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affective_state.R
\name{AffectiveState}
\alias{AffectiveState}
\title{AffectiveState}
\description{
Minimal structured affective layer with inertia-aware updates.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{state}}{A named list returned by \code{\link[=default_emotion_state]{default_emotion_state()}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AffectiveState-new}{\code{AffectiveState$new()}}
\item \href{#method-AffectiveState-decay}{\code{AffectiveState$decay()}}
\item \href{#method-AffectiveState-update_primary}{\code{AffectiveState$update_primary()}}
\item \href{#method-AffectiveState-describe}{\code{AffectiveState$describe()}}
\item \href{#method-AffectiveState-as_list}{\code{AffectiveState$as_list()}}
\item \href{#method-AffectiveState-clone}{\code{AffectiveState$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AffectiveState-new"></a>}}
\if{latex}{\out{\hypertarget{method-AffectiveState-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AffectiveState$new(state = default_emotion_state())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{state}}{Affective state used by \verb{$initialize()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AffectiveState-decay"></a>}}
\if{latex}{\out{\hypertarget{method-AffectiveState-decay}{}}}
\subsection{Method \code{decay()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AffectiveState$decay(current_time = Sys.time())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{current_time}}{Reference time used by \verb{$decay()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AffectiveState-update_primary"></a>}}
\if{latex}{\out{\hypertarget{method-AffectiveState-update_primary}{}}}
\subsection{Method \code{update_primary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AffectiveState$update_primary(updates)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{updates}}{Named numeric updates used by \verb{$update_primary()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AffectiveState-describe"></a>}}
\if{latex}{\out{\hypertarget{method-AffectiveState-describe}{}}}
\subsection{Method \code{describe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AffectiveState$describe(
  threshold = 0.2,
  include_blended = TRUE,
  method = "geometric"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{threshold}}{Threshold used by \verb{$describe()}.}

\item{\code{include_blended}}{Logical flag used by \verb{$describe()}.}

\item{\code{method}}{Combination method used by \verb{$describe()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AffectiveState-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-AffectiveState-as_list}{}}}
\subsection{Method \code{as_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AffectiveState$as_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AffectiveState-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AffectiveState-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AffectiveState$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
