% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{backup_agent}
\alias{backup_agent}
\title{Backup an \code{agentr} object with a timestamped filename}
\usage{
backup_agent(agent, dir = "agent_backups")
}
\arguments{
\item{agent}{An object created by \code{agentr}.}

\item{dir}{Backup directory.}
}
\value{
Invisibly returns the backup file path.
}
\description{
Saves a timestamped backup of an \code{agentr} core object to a specified
directory.
}
