% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affective_state.R
\name{decay_emotion_state}
\alias{decay_emotion_state}
\title{Apply time-based decay to an affective state}
\usage{
decay_emotion_state(emotion_state, current_time = Sys.time())
}
\arguments{
\item{emotion_state}{State list created by \code{\link[=default_emotion_state]{default_emotion_state()}} or
\code{\link[=define_random_emotion_state]{define_random_emotion_state()}}.}

\item{current_time}{Reference time.}
}
\value{
Updated affective state list.
}
\description{
Apply time-based decay to an affective state
}
