% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affective_state.R
\name{define_random_emotion_state}
\alias{define_random_emotion_state}
\title{Create a randomized affective state}
\usage{
define_random_emotion_state(
  total_intensity = 1,
  sparsity = 0,
  decay_rate = 0.98,
  inertia = 0.85
)
}
\arguments{
\item{total_intensity}{Total sum of primary emotion values.}

\item{sparsity}{Proportion of primary emotions to zero out.}

\item{decay_rate}{Hourly decay rate between 0 and 1.}

\item{inertia}{Inertia factor between 0 and 1 for incremental updates.}
}
\value{
A named list.
}
\description{
Create a randomized affective state
}
