% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffolder_message.R
\name{.normalize_candidate_values}
\alias{.normalize_candidate_values}
\title{Normalize candidate values}
\usage{
.normalize_candidate_values(x)
}
\arguments{
\item{x}{Candidate container.}
}
\value{
Character vector.
}
\description{
Normalize candidate values
}
\keyword{internal}
