% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffolder_message.R
\name{.predict_added_node_ids}
\alias{.predict_added_node_ids}
\title{Predict node ids generated by apply_human_feedback additions}
\usage{
.predict_added_node_ids(scaffolder, add)
}
\arguments{
\item{scaffolder}{A \code{\link{Scaffolder}} instance.}

\item{add}{Additions passed to \code{apply_human_feedback}.}
}
\value{
Character vector of predicted node ids.
}
\description{
Predict node ids generated by apply_human_feedback additions
}
\keyword{internal}
