% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_serialization.R
\name{.safe_save_json}
\alias{.safe_save_json}
\title{Save a JSON file safely}
\usage{
.safe_save_json(object, path, wait = 5, max_attempts = 10)
}
\arguments{
\item{object}{Object to serialize to JSON.}

\item{path}{Path to a JSON file.}

\item{wait}{Total wait time in seconds before giving up.}

\item{max_attempts}{Number of retry attempts.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Save a JSON file safely
}
\keyword{internal}
