% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_spec.R
\name{IACConfig}
\alias{IACConfig}
\title{IACConfig}
\description{
IACConfig

IACConfig
}
\details{
Configuration for interaction and communication.
}
\section{Methods}{

\describe{
\item{\verb{$initialize(enabled = TRUE, channels = character(), structured_io = TRUE, metadata = list())}}{Create an interaction and communication config.}
\item{\verb{$validate()}}{Validate the config.}
\item{\verb{$as_list()}}{Return a serializable representation.}
}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{enabled}}{Whether the subsystem is enabled.}

\item{\code{channels}}{Character vector of communication channels.}

\item{\code{structured_io}}{Whether strongly structured I/O is required.}

\item{\code{metadata}}{Free-form metadata list.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IACConfig-new}{\code{IACConfig$new()}}
\item \href{#method-IACConfig-validate}{\code{IACConfig$validate()}}
\item \href{#method-IACConfig-as_list}{\code{IACConfig$as_list()}}
\item \href{#method-IACConfig-clone}{\code{IACConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IACConfig-new"></a>}}
\if{latex}{\out{\hypertarget{method-IACConfig-new}{}}}
\subsection{Method \code{new()}}{
Create an interaction and communication config.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IACConfig$new(
  enabled = TRUE,
  channels = character(),
  structured_io = TRUE,
  metadata = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{enabled}}{Whether the subsystem is enabled.}

\item{\code{channels}}{Character vector of communication channels.}

\item{\code{structured_io}}{Whether strongly structured I/O is required.}

\item{\code{metadata}}{Free-form metadata list.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IACConfig-validate"></a>}}
\if{latex}{\out{\hypertarget{method-IACConfig-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the config.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IACConfig$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IACConfig-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-IACConfig-as_list}{}}}
\subsection{Method \code{as_list()}}{
Return a serializable representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IACConfig$as_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IACConfig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IACConfig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IACConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
