% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tool_config.R
\name{inferencer_available}
\alias{inferencer_available}
\title{Check whether \code{inferencer} is available}
\usage{
inferencer_available()
}
\value{
Logical scalar.
}
\description{
Check whether \code{inferencer} is available
}
