% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_graph_proposal.R
\name{KnowledgeGraphProposalState}
\alias{KnowledgeGraphProposalState}
\title{KnowledgeGraphProposalState}
\description{
KnowledgeGraphProposalState

KnowledgeGraphProposalState
}
\details{
State container for approved graph knowledge plus candidate proposals.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{approved_graph}}{Approved \code{agentr_knowledge_graph_spec}.}

\item{\code{proposals}}{Named list of \code{\link{KnowledgeGraphProposal}} objects.}

\item{\code{history}}{Proposal-state history.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KnowledgeGraphProposalState-new}{\code{KnowledgeGraphProposalState$new()}}
\item \href{#method-KnowledgeGraphProposalState-validate}{\code{KnowledgeGraphProposalState$validate()}}
\item \href{#method-KnowledgeGraphProposalState-add_proposal}{\code{KnowledgeGraphProposalState$add_proposal()}}
\item \href{#method-KnowledgeGraphProposalState-get_proposal}{\code{KnowledgeGraphProposalState$get_proposal()}}
\item \href{#method-KnowledgeGraphProposalState-list_proposals}{\code{KnowledgeGraphProposalState$list_proposals()}}
\item \href{#method-KnowledgeGraphProposalState-discuss_proposal}{\code{KnowledgeGraphProposalState$discuss_proposal()}}
\item \href{#method-KnowledgeGraphProposalState-approve_proposal}{\code{KnowledgeGraphProposalState$approve_proposal()}}
\item \href{#method-KnowledgeGraphProposalState-reject_proposal}{\code{KnowledgeGraphProposalState$reject_proposal()}}
\item \href{#method-KnowledgeGraphProposalState-approved_spec}{\code{KnowledgeGraphProposalState$approved_spec()}}
\item \href{#method-KnowledgeGraphProposalState-as_list}{\code{KnowledgeGraphProposalState$as_list()}}
\item \href{#method-KnowledgeGraphProposalState-clone}{\code{KnowledgeGraphProposalState$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeGraphProposalState-new"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeGraphProposalState-new}{}}}
\subsection{Method \code{new()}}{
Create a knowledge graph proposal state container.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeGraphProposalState$new(
  approved_graph = new_knowledge_graph_spec(),
  proposals = list(),
  history = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{approved_graph}}{Approved \code{agentr_knowledge_graph_spec} or serializable graph list.}

\item{\code{proposals}}{Initial proposals.}

\item{\code{history}}{Proposal-state history.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeGraphProposalState-validate"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeGraphProposalState-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the state object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeGraphProposalState$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeGraphProposalState-add_proposal"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeGraphProposalState-add_proposal}{}}}
\subsection{Method \code{add_proposal()}}{
Add a proposal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeGraphProposalState$add_proposal(proposal)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{proposal}}{\code{\link{KnowledgeGraphProposal}} object or serializable proposal record.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeGraphProposalState-get_proposal"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeGraphProposalState-get_proposal}{}}}
\subsection{Method \code{get_proposal()}}{
Return one proposal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeGraphProposalState$get_proposal(proposal_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{proposal_id}}{Proposal identifier.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeGraphProposalState-list_proposals"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeGraphProposalState-list_proposals}{}}}
\subsection{Method \code{list_proposals()}}{
List proposals with optional status filtering.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeGraphProposalState$list_proposals(status = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{status}}{Optional status filter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeGraphProposalState-discuss_proposal"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeGraphProposalState-discuss_proposal}{}}}
\subsection{Method \code{discuss_proposal()}}{
Discuss one proposal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeGraphProposalState$discuss_proposal(
  proposal_id,
  note,
  source = "human"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{proposal_id}}{Proposal identifier.}

\item{\code{note}}{Discussion or transition note.}

\item{\code{source}}{Discussion source.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeGraphProposalState-approve_proposal"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeGraphProposalState-approve_proposal}{}}}
\subsection{Method \code{approve_proposal()}}{
Approve one proposal and replace the approved graph.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeGraphProposalState$approve_proposal(proposal_id, note = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{proposal_id}}{Proposal identifier.}

\item{\code{note}}{Discussion or transition note.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeGraphProposalState-reject_proposal"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeGraphProposalState-reject_proposal}{}}}
\subsection{Method \code{reject_proposal()}}{
Reject one proposal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeGraphProposalState$reject_proposal(proposal_id, note = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{proposal_id}}{Proposal identifier.}

\item{\code{note}}{Discussion or transition note.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeGraphProposalState-approved_spec"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeGraphProposalState-approved_spec}{}}}
\subsection{Method \code{approved_spec()}}{
Return the approved graph.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeGraphProposalState$approved_spec()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeGraphProposalState-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeGraphProposalState-as_list}{}}}
\subsection{Method \code{as_list()}}{
Return a serializable representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeGraphProposalState$as_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeGraphProposalState-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeGraphProposalState-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeGraphProposalState$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
