% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_proposal_state.R
\name{KnowledgeProposalState}
\alias{KnowledgeProposalState}
\title{KnowledgeProposalState}
\description{
KnowledgeProposalState

KnowledgeProposalState
}
\details{
State container for approved knowledge plus active and historical proposals.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{approved_knowledge_spec}}{Approved \code{\link{KnowledgeSpec}}.}

\item{\code{proposals}}{Named list of \code{\link{KnowledgeProposal}} objects.}

\item{\code{history}}{Proposal-state history.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KnowledgeProposalState-new}{\code{KnowledgeProposalState$new()}}
\item \href{#method-KnowledgeProposalState-validate}{\code{KnowledgeProposalState$validate()}}
\item \href{#method-KnowledgeProposalState-add_proposal}{\code{KnowledgeProposalState$add_proposal()}}
\item \href{#method-KnowledgeProposalState-get_proposal}{\code{KnowledgeProposalState$get_proposal()}}
\item \href{#method-KnowledgeProposalState-list_proposals}{\code{KnowledgeProposalState$list_proposals()}}
\item \href{#method-KnowledgeProposalState-discuss_proposal}{\code{KnowledgeProposalState$discuss_proposal()}}
\item \href{#method-KnowledgeProposalState-approve_proposal}{\code{KnowledgeProposalState$approve_proposal()}}
\item \href{#method-KnowledgeProposalState-reject_proposal}{\code{KnowledgeProposalState$reject_proposal()}}
\item \href{#method-KnowledgeProposalState-approved_spec}{\code{KnowledgeProposalState$approved_spec()}}
\item \href{#method-KnowledgeProposalState-as_list}{\code{KnowledgeProposalState$as_list()}}
\item \href{#method-KnowledgeProposalState-clone}{\code{KnowledgeProposalState$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeProposalState-new"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeProposalState-new}{}}}
\subsection{Method \code{new()}}{
Create a knowledge proposal state container.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeProposalState$new(
  approved_knowledge_spec = KnowledgeSpec$new(),
  proposals = list(),
  history = list()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeProposalState-validate"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeProposalState-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the state object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeProposalState$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeProposalState-add_proposal"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeProposalState-add_proposal}{}}}
\subsection{Method \code{add_proposal()}}{
Add a proposal object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeProposalState$add_proposal(proposal)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeProposalState-get_proposal"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeProposalState-get_proposal}{}}}
\subsection{Method \code{get_proposal()}}{
Return one stored proposal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeProposalState$get_proposal(proposal_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeProposalState-list_proposals"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeProposalState-list_proposals}{}}}
\subsection{Method \code{list_proposals()}}{
List proposals with optional status filtering.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeProposalState$list_proposals(status = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeProposalState-discuss_proposal"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeProposalState-discuss_proposal}{}}}
\subsection{Method \code{discuss_proposal()}}{
Append a discussion note to one proposal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeProposalState$discuss_proposal(proposal_id, note, source = "human")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeProposalState-approve_proposal"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeProposalState-approve_proposal}{}}}
\subsection{Method \code{approve_proposal()}}{
Approve one proposal and add its item to approved knowledge.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeProposalState$approve_proposal(proposal_id, note = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeProposalState-reject_proposal"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeProposalState-reject_proposal}{}}}
\subsection{Method \code{reject_proposal()}}{
Reject one proposal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeProposalState$reject_proposal(proposal_id, note = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeProposalState-approved_spec"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeProposalState-approved_spec}{}}}
\subsection{Method \code{approved_spec()}}{
Return the approved knowledge specification.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeProposalState$approved_spec()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeProposalState-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeProposalState-as_list}{}}}
\subsection{Method \code{as_list()}}{
Return a serializable representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeProposalState$as_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeProposalState-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeProposalState-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeProposalState$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
