% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{load_agent}
\alias{load_agent}
\title{Load an \code{agentr} object from a file}
\usage{
load_agent(file_path)
}
\arguments{
\item{file_path}{File path from which to load the object.}
}
\value{
An object created by \code{agentr}.
}
\description{
Loads an \code{agentr} core object from a saved \code{.rds} file.
}
