% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{load_agent_spec}
\alias{load_agent_spec}
\title{Load an \code{AgentSpec} from a file}
\usage{
load_agent_spec(file_path)
}
\arguments{
\item{file_path}{File path from which to load the object.}
}
\value{
An \code{\link{AgentSpec}} object.
}
\description{
Loads a saved \code{\link{AgentSpec}} object from an \code{.rds} file.
}
