% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_review_spec.R
\name{load_design_review_spec}
\alias{load_design_review_spec}
\title{Load a design-review specification}
\usage{
load_design_review_spec(path)
}
\arguments{
\item{path}{Input \code{.rds} path.}
}
\value{
A \code{\link{DesignReviewSpec}} object.
}
\description{
Load a design-review specification
}
