% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_serialization.R
\name{load_json_file}
\alias{load_json_file}
\title{Load a JSON file}
\usage{
load_json_file(path, simplifyVector = TRUE)
}
\arguments{
\item{path}{Path to a JSON file.}

\item{simplifyVector}{Passed to \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}.}
}
\value{
Parsed JSON content.
}
\description{
Load a JSON file
}
