% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{load_knowledge_graph_spec_json}
\alias{load_knowledge_graph_spec_json}
\title{Load a knowledge graph specification from JSON}
\usage{
load_knowledge_graph_spec_json(file_path)
}
\arguments{
\item{file_path}{File path from which to load the JSON.}
}
\value{
An \code{agentr_knowledge_graph_spec} object.
}
\description{
Load a knowledge graph specification from JSON
}
