% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{load_knowledge_graph_spec_yaml}
\alias{load_knowledge_graph_spec_yaml}
\title{Load a knowledge graph specification from YAML}
\usage{
load_knowledge_graph_spec_yaml(file_path)
}
\arguments{
\item{file_path}{File path from which to load the YAML.}
}
\value{
An \code{agentr_knowledge_graph_spec} object.
}
\description{
Load a knowledge graph specification from YAML
}
