% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_proposal_state.R
\name{load_knowledge_proposal}
\alias{load_knowledge_proposal}
\title{Load a knowledge proposal}
\usage{
load_knowledge_proposal(path)
}
\arguments{
\item{path}{File path.}
}
\value{
A \code{\link{KnowledgeProposal}} object.
}
\description{
Load a knowledge proposal
}
