% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_proposal_state.R
\name{load_knowledge_spec}
\alias{load_knowledge_spec}
\title{Load a knowledge specification}
\usage{
load_knowledge_spec(path, format = c("rds", "json"))
}
\arguments{
\item{path}{File path.}

\item{format}{File format, either \code{rds} or \code{json}.}
}
\value{
A \code{\link{KnowledgeSpec}} object.
}
\description{
Load a knowledge specification
}
