% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_proposal_state.R
\name{load_knowledge_spec_json}
\alias{load_knowledge_spec_json}
\title{Load a knowledge specification from JSON}
\usage{
load_knowledge_spec_json(path)
}
\arguments{
\item{path}{File path.}
}
\value{
A \code{\link{KnowledgeSpec}} object.
}
\description{
Load a knowledge specification from JSON
}
