% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{load_memory_spec_json}
\alias{load_memory_spec_json}
\title{Load a \code{MemorySpec} from JSON}
\usage{
load_memory_spec_json(file_path)
}
\arguments{
\item{file_path}{File path from which to load the JSON.}
}
\value{
A \code{\link{MemorySpec}} object.
}
\description{
Load a \code{MemorySpec} from JSON
}
