% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{load_memory_spec_yaml}
\alias{load_memory_spec_yaml}
\title{Load a \code{MemorySpec} from YAML}
\usage{
load_memory_spec_yaml(file_path)
}
\arguments{
\item{file_path}{File path from which to load the YAML.}
}
\value{
A \code{\link{MemorySpec}} object.
}
\description{
Load a \code{MemorySpec} from YAML
}
