% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffolder_proposals.R
\name{load_workflow_proposal}
\alias{load_workflow_proposal}
\title{Load a workflow proposal}
\usage{
load_workflow_proposal(file_path)
}
\arguments{
\item{file_path}{File path from which to load the proposal.}
}
\value{
Workflow proposal object.
}
\description{
Loads a previously saved workflow proposal from an \code{.rds} file.
}
