% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_spec.R
\name{load_workflow_spec_json}
\alias{load_workflow_spec_json}
\title{Load a workflow specification from JSON}
\usage{
load_workflow_spec_json(file_path)
}
\arguments{
\item{file_path}{File path from which to load the workflow JSON.}
}
\value{
Workflow specification.
}
\description{
Load a workflow specification from JSON
}
