% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_spec.R
\name{load_workflow_spec_yaml}
\alias{load_workflow_spec_yaml}
\title{Load a workflow specification from YAML}
\usage{
load_workflow_spec_yaml(file_path)
}
\arguments{
\item{file_path}{File path from which to load the workflow YAML.}
}
\value{
Workflow specification.
}
\description{
Load a workflow specification from YAML
}
