% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_serialization.R
\name{load_yaml_file}
\alias{load_yaml_file}
\title{Load a YAML file}
\usage{
load_yaml_file(path)
}
\arguments{
\item{path}{Path to a YAML file.}
}
\value{
Parsed YAML content.
}
\description{
Load a YAML file
}
