% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_spec.R
\name{memory_persistence_policies}
\alias{memory_persistence_policies}
\title{Memory persistence policies}
\usage{
memory_persistence_policies()
}
\value{
Character vector of supported memory persistence policies.
}
\description{
Memory persistence policies
}
