% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_proposal.R
\name{MemoryProposalState}
\alias{MemoryProposalState}
\title{MemoryProposalState}
\description{
MemoryProposalState

MemoryProposalState
}
\details{
State container for approved memory schema plus candidate proposals.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{approved_memory_spec}}{Approved \code{\link{MemorySpec}}.}

\item{\code{proposals}}{Named list of \code{\link{MemoryProposal}} objects.}

\item{\code{history}}{Proposal-state history.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MemoryProposalState-new}{\code{MemoryProposalState$new()}}
\item \href{#method-MemoryProposalState-validate}{\code{MemoryProposalState$validate()}}
\item \href{#method-MemoryProposalState-add_proposal}{\code{MemoryProposalState$add_proposal()}}
\item \href{#method-MemoryProposalState-get_proposal}{\code{MemoryProposalState$get_proposal()}}
\item \href{#method-MemoryProposalState-list_proposals}{\code{MemoryProposalState$list_proposals()}}
\item \href{#method-MemoryProposalState-discuss_proposal}{\code{MemoryProposalState$discuss_proposal()}}
\item \href{#method-MemoryProposalState-approve_proposal}{\code{MemoryProposalState$approve_proposal()}}
\item \href{#method-MemoryProposalState-reject_proposal}{\code{MemoryProposalState$reject_proposal()}}
\item \href{#method-MemoryProposalState-approved_spec}{\code{MemoryProposalState$approved_spec()}}
\item \href{#method-MemoryProposalState-as_list}{\code{MemoryProposalState$as_list()}}
\item \href{#method-MemoryProposalState-clone}{\code{MemoryProposalState$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemoryProposalState-new"></a>}}
\if{latex}{\out{\hypertarget{method-MemoryProposalState-new}{}}}
\subsection{Method \code{new()}}{
Create a memory proposal state container.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MemoryProposalState$new(
  approved_memory_spec = MemorySpec$new(),
  proposals = list(),
  history = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{approved_memory_spec}}{Approved \code{\link{MemorySpec}} or serializable memory-spec list.}

\item{\code{proposals}}{Initial proposals.}

\item{\code{history}}{Proposal-state history.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemoryProposalState-validate"></a>}}
\if{latex}{\out{\hypertarget{method-MemoryProposalState-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the state object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MemoryProposalState$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemoryProposalState-add_proposal"></a>}}
\if{latex}{\out{\hypertarget{method-MemoryProposalState-add_proposal}{}}}
\subsection{Method \code{add_proposal()}}{
Add a proposal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MemoryProposalState$add_proposal(proposal)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{proposal}}{\code{\link{MemoryProposal}} object or serializable proposal record.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemoryProposalState-get_proposal"></a>}}
\if{latex}{\out{\hypertarget{method-MemoryProposalState-get_proposal}{}}}
\subsection{Method \code{get_proposal()}}{
Return one proposal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MemoryProposalState$get_proposal(proposal_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{proposal_id}}{Proposal identifier.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemoryProposalState-list_proposals"></a>}}
\if{latex}{\out{\hypertarget{method-MemoryProposalState-list_proposals}{}}}
\subsection{Method \code{list_proposals()}}{
List proposals with optional status filtering.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MemoryProposalState$list_proposals(status = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{status}}{Optional status filter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemoryProposalState-discuss_proposal"></a>}}
\if{latex}{\out{\hypertarget{method-MemoryProposalState-discuss_proposal}{}}}
\subsection{Method \code{discuss_proposal()}}{
Discuss one proposal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MemoryProposalState$discuss_proposal(proposal_id, note, source = "human")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{proposal_id}}{Proposal identifier.}

\item{\code{note}}{Discussion or transition note.}

\item{\code{source}}{Discussion source.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemoryProposalState-approve_proposal"></a>}}
\if{latex}{\out{\hypertarget{method-MemoryProposalState-approve_proposal}{}}}
\subsection{Method \code{approve_proposal()}}{
Approve one proposal and replace the approved memory spec.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MemoryProposalState$approve_proposal(proposal_id, note = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{proposal_id}}{Proposal identifier.}

\item{\code{note}}{Discussion or transition note.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemoryProposalState-reject_proposal"></a>}}
\if{latex}{\out{\hypertarget{method-MemoryProposalState-reject_proposal}{}}}
\subsection{Method \code{reject_proposal()}}{
Reject one proposal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MemoryProposalState$reject_proposal(proposal_id, note = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{proposal_id}}{Proposal identifier.}

\item{\code{note}}{Discussion or transition note.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemoryProposalState-approved_spec"></a>}}
\if{latex}{\out{\hypertarget{method-MemoryProposalState-approved_spec}{}}}
\subsection{Method \code{approved_spec()}}{
Return the approved memory specification.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MemoryProposalState$approved_spec()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemoryProposalState-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-MemoryProposalState-as_list}{}}}
\subsection{Method \code{as_list()}}{
Return a serializable representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MemoryProposalState$as_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemoryProposalState-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MemoryProposalState-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MemoryProposalState$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
