% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_graph.R
\name{new_knowledge_graph_spec}
\alias{new_knowledge_graph_spec}
\title{Create a knowledge-graph specification}
\usage{
new_knowledge_graph_spec(
  nodes = .empty_knowledge_graph_nodes(),
  edges = .empty_knowledge_graph_edges(),
  metadata = list()
)
}
\arguments{
\item{nodes}{Data frame of graph nodes.}

\item{edges}{Data frame of graph edges.}

\item{metadata}{Additional metadata list.}
}
\value{
An object of class \code{agentr_knowledge_graph_spec}.
}
\description{
Create a knowledge-graph specification
}
