% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_spec.R
\name{PGConfig}
\alias{PGConfig}
\title{PGConfig}
\description{
PGConfig

PGConfig
}
\details{
Configuration for the planning and goal-management subsystem.
}
\section{Methods}{

\describe{
\item{\verb{$initialize(enabled = TRUE, planning_mode = "task_decomposition", decomposition_style = "dag", metadata = list())}}{Create a planning and goal-management config.}
\item{\verb{$validate()}}{Validate the config.}
\item{\verb{$as_list()}}{Return a serializable representation.}
}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{enabled}}{Whether the subsystem is enabled.}

\item{\code{planning_mode}}{Planning mode label.}

\item{\code{decomposition_style}}{Workflow decomposition style.}

\item{\code{metadata}}{Free-form metadata list.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PGConfig-new}{\code{PGConfig$new()}}
\item \href{#method-PGConfig-validate}{\code{PGConfig$validate()}}
\item \href{#method-PGConfig-as_list}{\code{PGConfig$as_list()}}
\item \href{#method-PGConfig-print}{\code{PGConfig$print()}}
\item \href{#method-PGConfig-clone}{\code{PGConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGConfig-new"></a>}}
\if{latex}{\out{\hypertarget{method-PGConfig-new}{}}}
\subsection{Method \code{new()}}{
Create a planning and goal-management config.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGConfig$new(
  enabled = TRUE,
  planning_mode = "task_decomposition",
  decomposition_style = "dag",
  metadata = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{enabled}}{Whether the subsystem is enabled.}

\item{\code{planning_mode}}{Planning mode label.}

\item{\code{decomposition_style}}{Workflow decomposition style.}

\item{\code{metadata}}{Free-form metadata list.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGConfig-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PGConfig-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the config.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGConfig$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGConfig-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-PGConfig-as_list}{}}}
\subsection{Method \code{as_list()}}{
Return a serializable representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGConfig$as_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGConfig-print"></a>}}
\if{latex}{\out{\hypertarget{method-PGConfig-print}{}}}
\subsection{Method \code{print()}}{
Print a compact config summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGConfig$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Unused print arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGConfig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PGConfig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
