% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_review_spec.R
\name{preview_design_feedback}
\alias{preview_design_feedback}
\title{Preview design feedback application}
\usage{
preview_design_feedback(x, feedback, review_spec = NULL)
}
\arguments{
\item{x}{A \code{\link{Scaffolder}} or design object.}

\item{feedback}{Feedback item or list of items.}

\item{review_spec}{Optional review spec used for target-id warnings.}
}
\value{
A non-mutating preview list.
}
\description{
Preview design feedback application
}
