% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_graph.R
\name{print.agentr_knowledge_graph_spec}
\alias{print.agentr_knowledge_graph_spec}
\title{Format a knowledge-graph specification}
\usage{
\method{print}{agentr_knowledge_graph_spec}(x, ...)
}
\arguments{
\item{x}{Knowledge-graph specification.}

\item{...}{Unused.}
}
\description{
Format a knowledge-graph specification
}
