% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffolder_proposals.R
\name{print.agentr_workflow_proposal}
\alias{print.agentr_workflow_proposal}
\title{Format a workflow proposal}
\usage{
\method{print}{agentr_workflow_proposal}(x, ...)
}
\arguments{
\item{x}{Workflow proposal object.}

\item{...}{Unused.}
}
\description{
Format a workflow proposal
}
