% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_spec.R
\name{print.agentr_workflow_spec}
\alias{print.agentr_workflow_spec}
\title{Format a workflow specification}
\usage{
\method{print}{agentr_workflow_spec}(x, ...)
}
\arguments{
\item{x}{Workflow specification.}

\item{...}{Unused.}
}
\description{
Format a workflow specification
}
