% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_trace.R
\name{read_decision_traces}
\alias{read_decision_traces}
\title{Read decision traces}
\usage{
read_decision_traces(path)
}
\arguments{
\item{path}{JSONL or RDS path.}
}
\value{
List of traces.
}
\description{
Read decision traces
}
