% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_trace.R
\name{read_reflection_traces}
\alias{read_reflection_traces}
\title{Read reflection traces}
\usage{
read_reflection_traces(path)
}
\arguments{
\item{path}{JSONL or RDS path.}
}
\value{
List of traces.
}
\description{
Read reflection traces
}
