% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal_scaffolding.R
\name{render_markdown_terminal}
\alias{render_markdown_terminal}
\title{Render markdown-like text for terminal output}
\usage{
render_markdown_terminal(txt)
}
\arguments{
\item{txt}{Character string.}
}
\value{
Rendered character string with ANSI styling.
}
\description{
Render markdown-like text for terminal output
}
