% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{save_agent}
\alias{save_agent}
\title{Save an \code{agentr} object to a file}
\usage{
save_agent(agent, file_path)
}
\arguments{
\item{agent}{An object created by \code{agentr}.}

\item{file_path}{File path where the object should be saved.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Saves an \code{\link{AgentCore}}, \code{\link{CognitiveState}}, \code{\link{AffectiveState}}, or
\code{\link{Scaffolder}} object to a specified \code{.rds} file.
}
