% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{save_agent_spec}
\alias{save_agent_spec}
\title{Save an \code{AgentSpec} to a file}
\usage{
save_agent_spec(spec, file_path)
}
\arguments{
\item{spec}{An \code{\link{AgentSpec}} object.}

\item{file_path}{File path where the object should be saved.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Saves an \code{\link{AgentSpec}} object to a specified \code{.rds} file.
}
