% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_review_spec.R
\name{save_design_review_spec}
\alias{save_design_review_spec}
\title{Save a design-review specification}
\usage{
save_design_review_spec(x, path)
}
\arguments{
\item{x}{A \code{\link{DesignReviewSpec}} object.}

\item{path}{Output \code{.rds} path.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Save a design-review specification
}
