% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{save_knowledge_graph_spec_yaml}
\alias{save_knowledge_graph_spec_yaml}
\title{Save a knowledge graph specification as YAML}
\usage{
save_knowledge_graph_spec_yaml(spec, file_path)
}
\arguments{
\item{spec}{An \code{agentr_knowledge_graph_spec} object.}

\item{file_path}{File path where the YAML should be saved.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Save a knowledge graph specification as YAML
}
