% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_proposal_state.R
\name{save_knowledge_proposal}
\alias{save_knowledge_proposal}
\title{Save a knowledge proposal}
\usage{
save_knowledge_proposal(x, path)
}
\arguments{
\item{x}{A \code{\link{KnowledgeProposal}} object.}

\item{path}{File path.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Save a knowledge proposal
}
