% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_proposal_state.R
\name{save_knowledge_spec_json}
\alias{save_knowledge_spec_json}
\title{Save a knowledge specification as JSON}
\usage{
save_knowledge_spec_json(x, path)
}
\arguments{
\item{x}{A \code{\link{KnowledgeSpec}} object.}

\item{path}{File path.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Save a knowledge specification as JSON
}
