% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{save_memory_spec}
\alias{save_memory_spec}
\title{Save a \code{MemorySpec} to a file}
\usage{
save_memory_spec(spec, file_path, format = c("rds", "json", "yaml"))
}
\arguments{
\item{spec}{A \code{\link{MemorySpec}} object.}

\item{file_path}{File path where the object should be saved.}

\item{format}{File format, either \code{rds}, \code{json}, or \code{yaml}.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Saves a \code{\link{MemorySpec}} object to a specified \code{.rds}, \code{.json}, or \code{.yaml}
file.
}
