% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{save_memory_spec_yaml}
\alias{save_memory_spec_yaml}
\title{Save a \code{MemorySpec} as YAML}
\usage{
save_memory_spec_yaml(spec, file_path)
}
\arguments{
\item{spec}{A \code{\link{MemorySpec}} object.}

\item{file_path}{File path where the YAML should be saved.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Save a \code{MemorySpec} as YAML
}
