% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_spec.R
\name{save_workflow_spec}
\alias{save_workflow_spec}
\title{Save a workflow specification}
\usage{
save_workflow_spec(workflow, file_path, format = c("rds", "json", "yaml"))
}
\arguments{
\item{workflow}{Workflow specification.}

\item{file_path}{File path where the workflow should be saved.}

\item{format}{File format, either \code{rds}, \code{json}, or \code{yaml}.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Save a workflow specification
}
