% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_spec.R
\name{save_workflow_spec_json}
\alias{save_workflow_spec_json}
\title{Save a workflow specification as JSON}
\usage{
save_workflow_spec_json(workflow, file_path)
}
\arguments{
\item{workflow}{Workflow specification.}

\item{file_path}{File path where the JSON should be saved.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Save a workflow specification as JSON
}
