% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_telegram.R
\name{send_image_TG}
\alias{send_image_TG}
\title{Send an Image to Telegram}
\usage{
send_image_TG(image_path, caption_text = NULL, config)
}
\arguments{
\item{image_path}{File path to the image to be uploaded.}

\item{caption_text}{Optional caption text for the image.}

\item{config}{A list containing \code{bot_token}, \code{chat_id}, and \code{parse_mode}.}
}
\value{
Invisibly returns the API response (as an \code{httr} object).
}
\description{
Sends a photo with an optional caption to a Telegram chat.
}
