% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal_scaffolding.R
\name{terminal_ask_node_complete}
\alias{terminal_ask_node_complete}
\title{Ask for workflow-node completeness in the terminal}
\usage{
terminal_ask_node_complete(scaffolder, node_id)
}
\arguments{
\item{scaffolder}{A \code{\link{Scaffolder}} instance.}

\item{node_id}{Workflow node identifier.}
}
\value{
The human response string.
}
\description{
Ask for workflow-node completeness in the terminal
}
