% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal_scaffolding.R
\name{terminal_ask_node_rule}
\alias{terminal_ask_node_rule}
\title{Ask for a node-specific rule in the terminal}
\usage{
terminal_ask_node_rule(scaffolder, node_id)
}
\arguments{
\item{scaffolder}{A \code{\link{Scaffolder}} instance.}

\item{node_id}{Workflow node identifier.}
}
\value{
A list containing the prompt and response.
}
\description{
Ask for a node-specific rule in the terminal
}
