% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal_scaffolding.R
\name{terminal_ask_workflow_changes}
\alias{terminal_ask_workflow_changes}
\title{Ask for workflow changes in the terminal}
\usage{
terminal_ask_workflow_changes(scaffolder)
}
\arguments{
\item{scaffolder}{A \code{\link{Scaffolder}} instance.}
}
\value{
A list containing the prompt and response.
}
\description{
Ask for workflow changes in the terminal
}
