% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal_scaffolding.R
\name{terminal_scaffold_input}
\alias{terminal_scaffold_input}
\title{Prompt for terminal input during scaffolding}
\usage{
terminal_scaffold_input(prompt)
}
\arguments{
\item{prompt}{Character string shown to the human.}
}
\value{
Character string entered by the user.
}
\description{
Prompt for terminal input during scaffolding
}
