% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_graph.R
\name{validate_knowledge_graph_spec}
\alias{validate_knowledge_graph_spec}
\title{Validate a knowledge-graph specification}
\usage{
validate_knowledge_graph_spec(x)
}
\arguments{
\item{x}{Knowledge-graph specification.}
}
\value{
The validated object, invisibly.
}
\description{
Validate a knowledge-graph specification
}
