% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_spec.R
\name{validate_knowledge_item}
\alias{validate_knowledge_item}
\title{Validate a knowledge specification item}
\usage{
validate_knowledge_item(item)
}
\arguments{
\item{item}{Knowledge-item list.}
}
\value{
Validated item, invisibly.
}
\description{
Validate a knowledge specification item
}
