% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_proposal_class.R
\name{validate_knowledge_proposal}
\alias{validate_knowledge_proposal}
\title{Validate a knowledge proposal}
\usage{
validate_knowledge_proposal(x)
}
\arguments{
\item{x}{Knowledge proposal record or object.}
}
\value{
Validated proposal, invisibly.
}
\description{
Validate a knowledge proposal
}
