% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_spec.R
\name{validate_knowledge_spec}
\alias{validate_knowledge_spec}
\title{Validate a knowledge specification}
\usage{
validate_knowledge_spec(x)
}
\arguments{
\item{x}{A \code{\link{KnowledgeSpec}} object.}
}
\value{
The validated object, invisibly.
}
\description{
Validate a knowledge specification
}
