% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_spec.R
\name{validate_memory_spec}
\alias{validate_memory_spec}
\title{Validate a MemorySpec}
\usage{
validate_memory_spec(x)
}
\arguments{
\item{x}{A \code{\link{MemorySpec}} object.}
}
\value{
The validated object, invisibly.
}
\description{
Validate a MemorySpec
}
