% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffolder_message.R
\name{validate_scaffolder_message}
\alias{validate_scaffolder_message}
\title{Validate a machine-readable scaffolder message}
\usage{
validate_scaffolder_message(x, allowed_methods = scaffolder_action_methods())
}
\arguments{
\item{x}{Parsed scaffolder message.}

\item{allowed_methods}{Character vector of allowed method names.}
}
\value{
The validated message, invisibly.
}
\description{
Validate a machine-readable scaffolder message
}
