% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_spec.R
\name{KnowledgeSpec}
\alias{KnowledgeSpec}
\title{KnowledgeSpec}
\description{
KnowledgeSpec

KnowledgeSpec
}
\details{
Curated domain and epistemic knowledge used to guide agent behavior.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{items}}{Named list of knowledge items.}

\item{\code{metadata}}{Free-form metadata list.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KnowledgeSpec-new}{\code{KnowledgeSpec$new()}}
\item \href{#method-KnowledgeSpec-add_item}{\code{KnowledgeSpec$add_item()}}
\item \href{#method-KnowledgeSpec-get_item}{\code{KnowledgeSpec$get_item()}}
\item \href{#method-KnowledgeSpec-list_items}{\code{KnowledgeSpec$list_items()}}
\item \href{#method-KnowledgeSpec-validate}{\code{KnowledgeSpec$validate()}}
\item \href{#method-KnowledgeSpec-to_list}{\code{KnowledgeSpec$to_list()}}
\item \href{#method-KnowledgeSpec-print}{\code{KnowledgeSpec$print()}}
\item \href{#method-KnowledgeSpec-clone}{\code{KnowledgeSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeSpec-new}{}}}
\subsection{Method \code{new()}}{
Create a knowledge specification.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeSpec$new(items = list(), metadata = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeSpec-add_item"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeSpec-add_item}{}}}
\subsection{Method \code{add_item()}}{
Add a knowledge item.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeSpec$add_item(item)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeSpec-get_item"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeSpec-get_item}{}}}
\subsection{Method \code{get_item()}}{
Return a knowledge item by id.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeSpec$get_item(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeSpec-list_items"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeSpec-list_items}{}}}
\subsection{Method \code{list_items()}}{
List knowledge items with optional filters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeSpec$list_items(type = NULL, domain = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeSpec-validate"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeSpec-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the knowledge specification.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeSpec$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeSpec-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeSpec-to_list}{}}}
\subsection{Method \code{to_list()}}{
Return a serializable representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeSpec$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeSpec-print"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeSpec-print}{}}}
\subsection{Method \code{print()}}{
Print a compact summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeSpec$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KnowledgeSpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KnowledgeSpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KnowledgeSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
