% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_io.R
\name{load_agent}
\alias{load_agent}
\title{Load an XAgent object from a file}
\usage{
load_agent(file_path)
}
\arguments{
\item{file_path}{A string specifying the file path from which to load the agent.}
}
\value{
An object of class \code{XAgent}.
}
\description{
This function loads an XAgent object from a saved \code{.rds} file.
}
