% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{load_knowledge_graph_spec}
\alias{load_knowledge_graph_spec}
\title{Load a knowledge graph specification from a file}
\usage{
load_knowledge_graph_spec(file_path)
}
\arguments{
\item{file_path}{File path from which to load the object.}
}
\value{
An \code{agentr_knowledge_graph_spec} object.
}
\description{
Loads a saved \code{agentr_knowledge_graph_spec} object from an \code{.rds} file.
}
