% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{load_knowledge_graph_spec}
\alias{load_knowledge_graph_spec}
\title{Load a knowledge graph specification from a file}
\usage{
load_knowledge_graph_spec(file_path, format = c("rds", "json", "yaml"))
}
\arguments{
\item{file_path}{File path from which to load the object.}

\item{format}{File format, either \code{rds}, \code{json}, or \code{yaml}.}
}
\value{
An \code{agentr_knowledge_graph_spec} object.
}
\description{
Loads a saved \code{agentr_knowledge_graph_spec} object from an \code{.rds}, \code{.json},
or \code{.yaml} file.
}
