% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{load_memory_spec}
\alias{load_memory_spec}
\title{Load a \code{MemorySpec} from a file}
\usage{
load_memory_spec(file_path)
}
\arguments{
\item{file_path}{File path from which to load the object.}
}
\value{
A \code{\link{MemorySpec}} object.
}
\description{
Loads a saved \code{\link{MemorySpec}} object from an \code{.rds} file.
}
