% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_spec.R
\name{load_workflow_spec}
\alias{load_workflow_spec}
\title{Load a workflow specification}
\usage{
load_workflow_spec(file_path, format = c("rds", "json"))
}
\arguments{
\item{file_path}{File path from which to load the workflow.}

\item{format}{File format, either \code{rds} or \code{json}.}
}
\value{
Workflow specification.
}
\description{
Load a workflow specification
}
