% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_io.R
\name{save_agent}
\alias{save_agent}
\title{Save an XAgent object to a file}
\usage{
save_agent(agent, file_path)
}
\arguments{
\item{agent}{An object of class \code{XAgent}.}

\item{file_path}{A string specifying the file path where the agent should be saved.}
}
\value{
No return value. The function is called for its side effect.
}
\description{
This function saves an XAgent object to a specified \code{.rds} file.
}
