% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{save_knowledge_graph_spec}
\alias{save_knowledge_graph_spec}
\title{Save a knowledge graph specification to a file}
\usage{
save_knowledge_graph_spec(spec, file_path, format = c("rds", "json"))
}
\arguments{
\item{spec}{An \code{agentr_knowledge_graph_spec} object.}

\item{file_path}{File path where the object should be saved.}

\item{format}{File format, either \code{rds} or \code{json}.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Saves an \code{agentr_knowledge_graph_spec} object to a specified \code{.rds} or
\code{.json} file.
}
