% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_spec.R
\name{save_workflow_spec}
\alias{save_workflow_spec}
\title{Save a workflow specification}
\usage{
save_workflow_spec(workflow, file_path)
}
\arguments{
\item{workflow}{Workflow specification.}

\item{file_path}{File path where the workflow should be saved.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Save a workflow specification
}
