% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_spec.R
\name{validate_workflow_spec}
\alias{validate_workflow_spec}
\title{Validate a workflow specification}
\usage{
validate_workflow_spec(x)
}
\arguments{
\item{x}{Workflow specification.}
}
\value{
The validated object, invisibly.
}
\description{
Validate a workflow specification
}
